requirements.txt
aiohttp==3.7.4.post0
git+https://github.com/Rapptz/discord.py
discord~=1.7.3
aiosignal==1.2.0
async-timeout==3.0.1
dnspython==1.16.0
python-dotenv==0.19.2
pandas
PyNaCl==1.3.0
psutil==5.9.0
requests==2.27.1
asyncpg~=0.25.0
APScheduler==3.9.1
yarl==1.7.2
typing-extensions==4.1.1